LayoutTests/platform/ios-wk2/imported/w3c/web-platform-tests/html/rendering/non-replaced-elements/the-fieldset-and-legend-elements/legend-grid-flex-multicol-expected.txt

12

1
2

1
2

1
2

1
2

1
2

FAIL flex assert_equals: offsetHeight expected 20 but got 40
FAIL inline-flex assert_equals: offsetHeight expected 20 but got 40
FAIL grid assert_equals: offsetHeight expected 20 but got 40
FAIL inline-grid assert_equals: offsetHeight expected 20 but got 40
PASS multicol
